+++ content_aware_timelapse/viderator/html_on_image/assets/__init__.py
@@ -1,4 +1,4 @@
-""" """
+"""Makes looking up asset locations easier in other code"""
 
 from pathlib import Path