+++ test/assets/__init__.py
@@ -41,8 +41,10 @@ IRL_VS_SOFTWARE_1: List[Path] = [
     _ASSETS_DIRECTORY / "bench_scene_in_focus.png",
 ]
 
+STREAM_SOFTWARE = _ASSETS_DIRECTORY / "stream_software.png"
+
 IRL_VS_SOFTWARE_2: List[Path] = [
-    _ASSETS_DIRECTORY / "stream_software.png",
+    STREAM_SOFTWARE,
     _ASSETS_DIRECTORY / "bench_scene_in_focus.png",
 ]
 
@@ -53,7 +55,7 @@ IRL_VS_SOFTWARE_3: List[Path] = [
 
 SORTED_STREAM_GENERIC_PATHS: List[Path] = [
     _ASSETS_DIRECTORY / "stream_software_obscured_1.png",
-    _ASSETS_DIRECTORY / "stream_software.png",
+    STREAM_SOFTWARE,
     _ASSETS_DIRECTORY / "stream_cad_1.png",
     _ASSETS_DIRECTORY / "bench_scene_in_focus.png",
 ]